Add `cfg_tempdir` option
This allows users to set their own, possibly more persistent temporary
directory. If the parameter is blank (default), the argument is set to
`None`, which is the default for `tempfile.NamedTemporaryFile`.

Storing temporary files somewhere other than `tmp` is slightly safer in
cases if, say, the computer shuts down abnormally and a note is saved
but the editor is still open (before this change, the note's content
would be lost since the `tmp` directory is wiped on restart).