urlencode posted note data
Fixes #54 (+ signs being dropped from created/updated notes)

Note data was being posted to the server as json data, but for some
reason (which is undocumented), SimpleNote seems to expect to do further
processing on the posted data apart from simply parsing as json. The
correct method for sending is to first `urllib.parse.quote` the string
containing a json dump of the note.

This was a regression introduced with the switch to using the requests
library in commit 4110c45. Unfortunately simplenote still accepted notes
without being quoted, silently dropping characters like the plus sign.