fix code execution on copy bug
Copying text from notes to the clipboard was being performed by building
a shell command to execute. This resulted in the line being copied
substituted directly into the shell command. A carefully crafted line
could run any arbitrary shell command, and some lines could crash the
process causing the line to fail to copy.

This fixes the issue by not using a shell to interpret the command, and
passing the text to be copied directly to stdin.