Fix tempfile suffix logic
Previously the file extension (suffix) only depended on
the value of `cfg_default_markdown`,
which wasn't helpful for editing existing notes.

Ensure that for existing notes,
the file extension for tempfiles passed to the editor
matches that of the note type (plain text or markdown).
Retain the old behaviour for new notes.